# docs/swagger.yaml
definitions:
  structs.Metric:
    properties:
      delta:
        description: значение метрики в случае передачи counter
        type: integer
      hash:
        description: hmac метрики
        type: string
      id:
        description: имя метрики
        type: string
      type:
        description: параметр, принимающий значение gauge или counter
        type: string
      value:
        description: значение метрики в случае передачи gauge
        type: number
    type: object
  structs.MetricGet:
    properties:
      id:
        description: имя метрики
        example: CPU
        type: string
      type:
        description: параметр, принимающий значение gauge или counter
        example: gauge
        type: string
    type: object
  structs.Response:
    properties:
      error:
        type: string
      hash:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: Service for storing and retreiving metrics
  title: Monitoring API
paths:
  /ping:
    get:
      description: Checking if DB is available
      produces:
      - application/json
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
      summary: Ping Database
      tags:
      - self-health
  /update/:
    post:
      description: Set or Update metrics value
      parameters:
      - description: Metric to set/update
        in: body
        name: metrics
        required: true
        schema:
          $ref: '#/definitions/structs.Metric'
      produces:
      - application/json
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Response'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/structs.Response'
      summary: Set/Update metric
      tags:
      - metrics
  /update/{metricType}/{metricID}/{metricValue}:
    post:
      description: Set or Update metric value
      parameters:
      - description: metric type
        enum:
        - counter
        - gauge
        in: path
        name: metricType
        required: true
        type: string
      - description: metric id
        in: path
        name: metricID
        required: true
        type: string
      - description: metric value
        in: path
        name: metricValue
        required: true
        type: string
      produces:
      - application/json
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Response'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/structs.Response'
      summary: Set/Update metric
      tags:
      - metrics
  /updates/:
    post:
      description: Set or Update multiple metrics at once
      parameters:
      - description: List of metrics to set/update
        in: body
        name: metrics
        required: true
        schema:
          items:
            $ref: '#/definitions/structs.Metric'
          type: array
      produces:
      - application/json
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Response'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/structs.Response'
      summary: Set/Update metrics
      tags:
      - metrics
  /value/:
    post:
      consumes:
      - application/json
      description: Retreiving metric value
      parameters:
      - description: Get value for metric
        in: body
        name: metric
        required: true
        schema:
          $ref: '#/definitions/structs.MetricGet'
      produces:
      - application/json
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Metric'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.Response'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/structs.Response'
      summary: Get  metric
      tags:
      - metrics
  /value/{metricType}/{metricID}:
    get:
      description: Retreiving metric value
      parameters:
      - description: metric type
        enum:
        - counter
        - gauge
        in: path
        name: metricType
        required: true
        type: string
      - description: metric id
        in: path
        name: metricID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Metric'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.Response'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/structs.Response'
      summary: Get metric
      tags:
      - metrics
swagger: "2.0"